WP Loader: Reduce the publish rate to 0.1 Hz
It is completely pointless to push the same list of 10k waypoints
through the pipes 40 times per second. The serialization and
deserialization costs are huge and it's impossible to even run the
whole thing reliably on weaker systems. 0.1 Hz is probably too much
anyways, but let's imagine we need to handle cases where the waypoints
actually change, like re-routing or similar.